% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{cdf.maf}
\alias{cdf.maf}
\title{Compute cumulative probability of mutant allele fraction (MAF)}
\usage{
cdf.maf(maf, n.read, maf.mtx, nreps = NULL)
}
\arguments{
\item{maf}{A vector of values at which to evaluate the CDF}

\item{n.read}{Number of reads observed in the cells}

\item{maf.mtx}{Mutant allele fraction matrix from maf.pmf}

\item{nreps}{Number of generations, default is column of maf.mtx minus 1}
}
\value{
The cumulative probability by given MAF
}
\description{
Compute cumulative probability of MAF for a population of cells by a given MAF distribution matrix
}
\examples{
set.seed(1234)
tot.read = 20
mut.read = sample(tot.read,15,replace = TRUE)
wt.read = tot.read - mut.read
maf.mtx = maf.pmf(mut.mito = 1,
          wt.mito = 19,
          nreps = 5,
          beta = rep(0,4)) ## example for the model without selection
maf = 0:tot.read/tot.read
cdf.maf(maf,n.read = tot.read,maf.mtx = maf.mtx)
}
\seealso{
\code{\link{maf.pmf}}
}
