% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{dcd.mtx}
\alias{dcd.mtx}
\title{Compute daughter cell mutant allele fraction distribution}
\usage{
dcd.mtx(n.mito, beta = rep(0, 4))
}
\arguments{
\item{n.mito}{Mitochondrial copy number of parent cell}

\item{beta}{Coefficients for cubic function for log.psi}
}
\value{
Matrix of daughter cell mutant allele fraction distribution
}
\description{
Compute daughter cell mutant allele fraction distribution
}
\examples{
maf.dcd = dcd.mtx(n.mito = 20,beta = rep(0,4)) ## example for the model without selection
}
