% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{full.mle}
\alias{full.mle}
\title{Estimation of selection and non-selection models}
\usage{
full.mle(
  mut.read,
  wt.read,
  nmito.start,
  nreps,
  selection.reps,
  selection.mut.start
)
}
\arguments{
\item{mut.read}{Initial mutant mitochondria read count for each single cell (vector)}

\item{wt.read}{Initial wildtype mitochondria read count for each single cell (vector)}

\item{nmito.start}{Mitochondrial copy number of parent cell (scalar)}

\item{nreps}{Number of generations (scalar)}

\item{selection.reps}{Number of generations for selection model estimation (vector between (0,nreps))}

\item{selection.mut.start}{Number of mutant mitochondrial read count for selection model estimation (vector between (0,nmito.start))}
}
\value{
A list of input (mut.read, wt.read, nmito.start, number of generations) and res.table
}
\description{
Calculate the negative log-likelihood of the model and estimate coefficients for cubic function for log.psi
}
\details{
The function calculates the negative log-likelihood of non-selection models of all numbers of mutant mitochondrial read count and all generations.
The function also calculates the negative log-likelihood of selection models with specified numbers of mutant mitochondrial read count and specified generation.
Coefficients for cubic function for log.psi (log odds ratio of mutant vs. wildtype read count) are estimated for selection model.
}
\examples{
set.seed(1234)
tot.read = 10                               ## total read count
mut.read = sample(tot.read,10,replace = TRUE)
wt.read = tot.read - mut.read
res = full.mle(mut.read = mut.read,
               wt.read = wt.read,
               nmito.start = 10,
               nreps = 5,

              ## specify the generation in selection model estimation
              selection.reps = c(3,5),

              ## specify the number of mutant mitochondrial read count in selection model estimation
              selection.mut.start = c(3,7))
}
\seealso{
\code{\link{nlogL.null}} \code{\link{local.mle}}
}
