% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{get.best.model}
\alias{get.best.model}
\title{Obtain the best selection model using clustering method}
\usage{
get.best.model(res.tbl, K = 10, show.best = TRUE)
}
\arguments{
\item{res.tbl}{res.table from {\link{full.mle}} function, containing negative log-likelihood values and estimated coefficients of each model}

\item{K}{Number of cluster}

\item{show.best}{A logical indicator. If TRUE then result shows the best model from all clusters. If FALSE then the result shows the best model of each cluster.}
}
\value{
A data frame with the best selection model of low/high-starting MAF and the best no selection model at the last row.
}
\description{
Clustering on the selection models and find the best one. The best selection model can be from low-starting mutant allele frequency (MAF)
model or high-starting MAF model or both
}
\details{
Use hierarchical clustering method to cluster selection models and find the best one. The best model of each cluster could be with high-staring MAF (MAF>50%) or with low-starting MAF (MAF<=50%).
}
\examples{
res.tbl.best = get.best.model(res.tbl = res.tbl,K = 10,show.best = TRUE)
}
\seealso{
\code{\link{full.mle}}
}
