% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hg.hist}
\alias{hg.hist}
\title{Calculate the histogram data for the mutant allele fraction in the mtDNA mutation model}
\usage{
hg.hist(read.data, result.tbl, nbin = 10000)
}
\arguments{
\item{read.data}{Read count matrix, one row per cell, first column with mutant counts, second with reference counts}

\item{result.tbl}{Table with columns: nmito.start, mutant.start, generation, beta0-beta3, clr(optional)}

\item{nbin}{Number of bin for histogram}
}
\value{
A list includes two numeric vectors: first with MAF point, second with adjusted mutant read frequency
}
\description{
Calculate the histogram data for the mutant allele fraction in the mtDNA mutation model
}
\examples{
res.tbl.best = get.best.model(res.tbl = res.tbl,K = 10,show.best = TRUE)
best.null = which(rowSums(res.tbl.best[,grep("beta",colnames(res.tbl.best))])==0)
null.hist = hg.hist(read.data = reads,                  ## example for the model without selection
                   result.tbl = res.tbl.best[best.null,],
                   nbin = 1000)
}
