% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hg.hist.smooth}
\alias{hg.hist.smooth}
\title{Calculate the beta-smoothed histogram data for the mutant allele fraction in the mtDNA mutation model}
\usage{
hg.hist.smooth(maf, read.data, mdl)
}
\arguments{
\item{maf}{Matrix of the distribution of mutant allele fraction for a population of cells}

\item{read.data}{A matrix with columns: mut.read and wt.read, representing observed mutant and wildtype mitochondria read count for each single cell}

\item{mdl}{Table with columns: nmito.start, mutant.start, generation, beta0-beta3, clr(optional)}
}
\value{
A vector of beta-smoothed mutant read frequency
}
\description{
Calculate the beta-smoothed histogram data for the mutant allele fraction in the mtDNA mutation model
}
\examples{
res.tbl.best = get.best.model(res.tbl = res.tbl,K = 10,show.best = TRUE)
best.null = which(rowSums(res.tbl.best[,grep("beta",colnames(res.tbl.best))])==0)
null.mdl = res.tbl[best.null,]
null.maf = maf.pmf(res.tbl.best$mutant.start[best.null],
                 res.tbl.best$nmito.start[best.null]-res.tbl.best$mutant.start[best.null],
                 res.tbl.best$generation[best.null])
smooth.hist = hg.hist.smooth(maf = null.maf,read.data = reads,mdl = null.mdl)
}
\seealso{
\code{\link{maf.pmf}}, \code{\link{read.pmf}}
}
