% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{local.mle}
\alias{local.mle}
\title{Estimation for selection and non-selection models of specified number of mutant/wildtype mitochondrial read count}
\usage{
local.mle(mut.read, wt.read, mut.start, wt.start, nreps, beta0 = rep(0, 4))
}
\arguments{
\item{mut.read}{Initial mutant mitochondria read count for each single cell (vector)}

\item{wt.read}{Initial wildtype mitochondria read count for each single cell (vector)}

\item{mut.start}{Number of mutant mitochondrial read count (scalar)}

\item{wt.start}{Number of mutant mitochondrial read count (scalar)}

\item{nreps}{Number of generations (scalar)}

\item{beta0}{Coefficients for cubic function for log.psi}
}
\value{
A list includes betas estimation for selection model, negative log-likelihood value of selection and non-selection model, model convergence information,and input values
}
\description{
Calculate negative log-likelihood of the model and estimate cubic function coefficients for log odds ratio in non-central hypergeometric distribution
}
\examples{
set.seed(1234)
tot.read = 10                               ## total read count
mut.read = sample(tot.read,10,replace = TRUE)
wt.read = tot.read - mut.read   ## the sum of mut.start and wt.start should be total read count
mle = local.mle(mut.read = mut.read,
                wt.read = wt.read,
                mut.start = 3,
                wt.start = 7,
                nreps = 5,
                beta0 = rep(0,4))
}
\seealso{
\code{\link{maf.pmf}},\code{\link{dcd.mtx}},\code{\link{read.pmf}},\code{\link{dnchg}},\code{\link{nlogL.maf}}
}
