% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{maf.pmf}
\alias{maf.pmf}
\title{Compute distribution of mutant allele fraction for a population of cells}
\usage{
maf.pmf(mut.mito, wt.mito, nreps, beta = rep(0, 4))
}
\arguments{
\item{mut.mito}{Number of mutant mitochondrial read count}

\item{wt.mito}{Number of wildtype mitochondrial read count}

\item{nreps}{Number of generations (scalar)}

\item{beta}{Coefficients for cubic function for log.psi}
}
\value{
Matrix of distribution of mutant allele fraction for a population of cells descending over generations
}
\description{
Compute distribution of mutant allele fraction for a population of cells descending over generations from an initiating cell
}
\examples{
## example for the model without selection
maf.mtx = maf.pmf(mut.mito = 1,
                  wt.mito = 19,
                  nreps = 5,
                  beta = rep(0,4))
}
