% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{nlogL.maf}
\alias{nlogL.maf}
\title{Calculate of negative log-likelihood of the model}
\usage{
nlogL.maf(beta, mut.read, wt.read, mut.start, wt.start, nreps)
}
\arguments{
\item{beta}{Coefficients for cubic function for log.psi}

\item{mut.read}{Initial mutant mitochondria read count for each single cell (vector)}

\item{wt.read}{Initial wildtype mitochondria read count for each single cell (vector)}

\item{mut.start}{Number of mutant mitochondrial read count (scalar)}

\item{wt.start}{Number of wildtype mitochondrial read count (scalar)}

\item{nreps}{Number of generations (scalar)}
}
\value{
Negative log-likelihood value
}
\description{
Calculate of negative log-likelihood of the model
}
\examples{
set.seed(1234)
tot.read = 10                       ## total read count
mut.read = sample(tot.read,10,replace = TRUE)
wt.read = tot.read - mut.read    ## the sum of mut.start and wt.start should be total read count
## example for the model without selectio
neg_ll = nlogL.maf(beta = rep(0,4),
                   mut.read = mut.read,
                   wt.read = wt.read,
                   mut.start = 3,
                   wt.start = 7,
                   nreps = 5)
}
\seealso{
\code{\link{maf.pmf}},\code{\link{dcd.mtx}},\code{\link{read.pmf}},\code{\link{dnchg}}
}
