% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{nlogL.null}
\alias{nlogL.null}
\title{Mitochondrial mutations prevalence modeling without selective pressure}
\usage{
nlogL.null(mut.read, wt.read, nmito.start = 100, nreps = 100)
}
\arguments{
\item{mut.read}{Initial mutant mitochondria read count for each single cell (vector)}

\item{wt.read}{Initial wildtype mitochondria read count for each single cell (vector)}

\item{nmito.start}{Mitochondrial copy number of parent cell (scalar)}

\item{nreps}{Number of generations (scalar)}
}
\value{
A list includes all inputs (mut.read,wt.read,nmito.start,nreps) and a data frame with calculated negative log-likelihood value of all models .
}
\description{
Calculate negative log-likelihood for model of mutant mitochondria read count (from 1 to total copy number) and generation (from 1 to specified generation)
}
\details{
The modeling without selective pressure of mutant mtDNA prevalence in subsequent generations is by iterative convolution of hypergeometric models, in which the mutant read is modeled by hypergeometric sampling for a given number of mutant/wildtype mitochondria read.
}
\examples{
tot.read = 20
mut.read = sample(tot.read,15,replace = TRUE)
wt.read = tot.read - mut.read
null.res = nlogL.null(mut.read,wt.read,nmito.start = 100,nreps = 50)
}
\seealso{
\code{\link{maf.pmf}},\code{\link{dcd.mtx}},\code{\link{read.pmf}},\code{\link{dnchg}}
}
