% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{P_value}
\alias{P_value}
\title{Compute P-value of likelihood ratio test between two models}
\usage{
P_value(
  res.tbl,
  nreps = NULL,
  mutant.start = NULL,
  baseline.model = "best-non-selection",
  alt.model = "selection"
)
}
\arguments{
\item{res.tbl}{res.table from {\link{full.mle}} function, containing negative log-likelihood values of each model}

\item{nreps}{Number of generations (scalar)}

\item{mutant.start}{Number of mutant mitochondrial read count (scalar)}

\item{baseline.model}{A character string specifying the null model, must be one of "best-non-selection"(default) or "best-selection"}

\item{alt.model}{A character string specifying the alternative model, must be one of "selection" (default) or "non-selection"}
}
\value{
A list includes inputs (specified nreps and mutant.start for models in comparison),the best selection and non-selection model,
the baseline model and res.table.pval table with p-values from LR tests
}
\description{
Compute P-value of likelihood ratio (LR) test between two models
}
\details{
The function calculates p-value of LR test of two models. E.g., for identifying selective pressure from mitochondrial DNA mutation, We assesses the goodness of fit of
selection model and the best non-selection by LR test.
}
\examples{
res = P_value(res.tbl = res.tbl,
                  nreps = NULL,
                  mutant.start = NULL,
                  baseline.model =  "best-non-selection",
                  alt.model = "selection")
}
\seealso{
\code{\link{full.mle}}
}
