% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.hist}
\alias{plt.hist}
\title{Plot histogram of mtDNA for the mutant allele fraction in the mtDNA mutation model}
\usage{
plt.hist(
  read.data,
  res.tbl.best,
  nbin = 1000,
  clr.scheme = "rainbow",
  show.logOR = TRUE,
  plot.title = NULL
)
}
\arguments{
\item{read.data}{A matrix with columns: mut.read and wt.read, representing observed mutant and wildtype mitochondria read count for each single cell}

\item{res.tbl.best}{A data frame from get.best.model() output with columns: nmito.start, mutant.start, generation, beta0-beta3, clr(optional), each row represents a model}

\item{nbin}{Number of bin for histogram}

\item{clr.scheme}{Default color scheme for models}

\item{show.logOR}{A logical indicator. If TRUE, the plot showes log odds ratio of mutant mtDNA over wildtype mtDNA of the best selection model}

\item{plot.title}{A character string, default is NULL}
}
\value{
Probability mass histogram for the mutant allele fraction in the observed read count of selection model and no selection model
}
\description{
Plot the histogram of the selection model, selection-free model and observed reads data
}
\examples{
res.tbl.best = get.best.model(res.tbl= res.tbl,K = 10,show.best = TRUE)
res.tbl.best$clr = c("red","green","grey")
plt.hist(read.data = reads,
          res.tbl.best = res.tbl.best,
          nbin = 1000,
          clr.scheme = "rainbow",
          show.logOR = TRUE,
          plot.title = NULL)
}
\seealso{
\code{\link{get.best.model}}, \code{\link{sc.hist}}, \code{\link{hg.hist}}
}
