% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.hist.smooth}
\alias{plt.hist.smooth}
\title{Plot beta-smoothed histogram of mtDNA for the mutant allele fraction in the mtDNA mutation model}
\usage{
plt.hist.smooth(read.data, res.tbl.best, plot.title = NULL)
}
\arguments{
\item{read.data}{A matrix with columns: mut.read and wt.read, representing observed mutant and wildtype mitochondria read count for each single cell}

\item{res.tbl.best}{A data frame from get.best.model() output with columns: nmito.start, mutant.start, generation, beta0-beta3, clr(optional), each row represents a model}

\item{plot.title}{A character string, default is NULL}
}
\value{
Beta-smoothed histogram
}
\description{
Plot beta-smoothed histogram of mtDNA for the mutant allele fraction in observed data and modeling
}
\examples{
res.tbl.best = get.best.model(res.tbl = res.tbl,K = 10,show.best = TRUE)
res.tbl.best$clr = c("red","green","grey")
plt.hist.smooth(read.data = reads,
                 res.tbl.best = res.tbl.best,
                 plot.title = NULL)
}
\seealso{
\code{\link{hg.hist.smooth}}
}
