% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.mtDNA.gens}
\alias{plt.mtDNA.gens}
\title{Plot the mtDNA mutation probability distribution over generations}
\usage{
plt.mtDNA.gens(res.tbl.best, ngen.show = 5, clr.scheme = "rainbow")
}
\arguments{
\item{res.tbl.best}{A data frame from get.best.model() output with columns: nmito.start, mutant.start, generation, beta0-beta3, clr(optional), each row represents a model}

\item{ngen.show}{The number of segment of generations to show}

\item{clr.scheme}{Default color scheme for models}
}
\value{
Plot of mtDNA mutation probability distribution over generations
}
\description{
Plot the mtDNA mutation probability distribution over generations
}
\examples{
res.tbl.best = get.best.model(res.tbl,K = 10,show.best = TRUE)
res.tbl.best$clr = c("red","green","grey")
plt.mtDNA.gens(res.tbl.best = res.tbl.best, ngen.show=5)
}
\seealso{
\code{\link{get.best.model}} to get the best selection model and best no selection model
}
