% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{read.pmf}
\alias{read.pmf}
\title{Probability distribution function for mutant read counts}
\usage{
read.pmf(mut.read, wt.read, maf.mtx, nreps = NULL)
}
\arguments{
\item{mut.read}{Mutant mitochondria read count for each single cell (vector)}

\item{wt.read}{Wildtype mitochondria read count for each single cell (vector)}

\item{maf.mtx}{Mutant allele fraction distribution matrix estimated from maf.pmf function}

\item{nreps}{Number of generations (scalar)}
}
\value{
A vector of probability of mutant read counts
}
\description{
Probability distribution function for mutant read counts
}
\examples{
## example for the model without selection
set.seed(1234)
maf.mtx = maf.pmf(mut.mito = 1,wt.mito = 19,nreps = 5,beta = rep(0,4))
tot.read = 10
mut.read = sample(tot.read,10,replace = TRUE)
wt.read = tot.read - mut.read
read.pmf(mut.read,wt.read,maf.mtx)
mut.all = 0:tot.read
wt.all = tot.read:0
sum(read.pmf(mut.all,wt.all,maf.mtx))
}
\seealso{
\code{\link{maf.pmf}}
}
