% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads}
\alias{reads}
\title{Single cell RNA-seq for 1888 cells with at least one mtDNA read}
\format{
A matrix with 1888 cells (rows) and reads in columns (mutant and wildtype)
}
\usage{
reads
}
\description{
An example single cell RNA-seq for 1888 cells with at least one mtDNA read from a diagnostic pediatric acute lymphoblastic leukemia sample
}
\keyword{datasets}
