% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sc.hist}
\alias{sc.hist}
\title{Calculate the histogram data for the mutant allele fraction in the observed data}
\usage{
sc.hist(read.data, nbin = 10000)
}
\arguments{
\item{read.data}{A matrix with columns: mut.read and wt.read, representing observed mutant and wildtype mitochondria read count for each single cell}

\item{nbin}{Number of bin for histogram}
}
\value{
A data.frame: first column with MAF point, second with adjusted mutant read frequency
}
\description{
Calculate the histogram data for the mutant allele fraction in the observed data
}
\examples{
obs.hist = sc.hist(read.data = reads, nbin = 1000)
}
